Address OpenTag migration review comments
