+++ .env.example
@@ -1,4 +1,4 @@
-# ── Canonical Intelligence Channel host (`pnpm channel`) ───────────────
+# ── Intelligence Channel host (`pnpm channel`) ─────────────────────────
 # Both variables are required. The local Python agent listens on :8123.
 AGENT_URL=http://localhost:8123/
 INTELLIGENCE_API_KEY=cpk-...
@@ -6,38 +6,32 @@ INTELLIGENCE_API_KEY=cpk-...
 # Production defaults; override only for another Intelligence environment.
 # INTELLIGENCE_API_URL=https://api.intelligence.copilotkit.ai
 # INTELLIGENCE_GATEWAY_WS_URL=wss://realtime.intelligence.copilotkit.ai
-# Defaults to "opentag". The production Railway service sets this to "kite"
-# so the existing Slack app keeps its @kite handle.
+# Base Channel name. OpenTag declares this Slack Channel plus a Teams Channel
+# suffixed with "-teams". Railway sets the base to "kite".
 # INTELLIGENCE_CHANNEL_NAME=opentag
 # HTTP port for the CopilotRuntime listener. Defaults to 3000.
 # PORT=3000
 # Optional authorization header forwarded to the AG-UI agent.
 # AGENT_AUTH_HEADER=Bearer ...
 
-# Attach managed Slack and Microsoft Teams integrations in Intelligence. The
-# optional credentials below are only for direct local-development adapters
-# inside this same Intelligence-owned runtime.
+# Slack and Microsoft Teams credentials belong in CopilotKit Intelligence, not
+# this runtime environment. Discord, Telegram, and WhatsApp are coming soon.
 
-# Slack Socket Mode: set both values or neither.
-# SLACK_BOT_TOKEN=xoxb-...
-# SLACK_APP_TOKEN=xapp-...
+# ── Python deep-research agent (`pnpm agent`) ───────────────────────────
+# The Python service also loads this root .env during local development.
+OPENAI_API_KEY=sk-...
+# Optional live web research. Without Tavily the agent still chats, plans,
+# uses its virtual filesystem, and renders UI components.
+# TAVILY_API_KEY=tvly-...
+# OPENAI_MODEL=gpt-5.5
+# SERVER_HOST=0.0.0.0
+# SERVER_PORT=8123
 
-# Microsoft Teams / Azure Bot: set both values or neither.
-# TEAMS_CLIENT_ID=...
-# TEAMS_CLIENT_SECRET=...
-# TEAMS_TENANT_ID=...
-# TEAMS_PORT=3978
+# ── Internal research sources (optional) ────────────────────────────────
+# LINEAR_API_KEY=lin_api_...
+# LINEAR_MCP_URL=https://mcp.linear.app/mcp
 
-# Discord, Telegram, and WhatsApp are coming soon. They have no active
-# configuration in this launch.
-
-# The Python service loads `agent/.env`; copy `agent/.env.example` there for
-# OPENAI_API_KEY and optional Tavily/Linear/Notion research variables.
-
-# Notion remains available for local development through the optional
-# `pnpm notion-mcp` sidecar. These root variables configure that sidecar; put
-# the same NOTION_MCP_AUTH_TOKEN in `agent/.env`. Notion is not a Railway
-# service or launch blocker.
+# The Notion MCP sidecar is local-only and not a Railway service.
 # NOTION_TOKEN=ntn_...
 # NOTION_MCP_AUTH_TOKEN=choose-a-strong-shared-secret
 # NOTION_MCP_URL=http://127.0.0.1:3001/mcp
