+++ README.md
@@ -8,7 +8,6 @@ The launch supports:
 
 - Slack through Intelligence, including the existing production `@kite` app.
 - Microsoft Teams through Intelligence.
-- Optional direct Slack and Teams adapters for local development.
 
 Discord, Telegram, and WhatsApp are coming soon.
 
@@ -31,12 +30,11 @@ agent (Python + LangGraph deepagents)
           └── Notion MCP (optional local sidecar)
 ```
 
-There is one canonical Channel host: [`server.ts`](./server.ts). It creates one
-`CopilotKitIntelligence`, one `CopilotRuntime`, and an adapter-free managed
-OpenTag Channel. Platform attachments are normally configured in Intelligence.
-Supplying local Slack or Teams credentials adds a separately named direct
-Channel to that same runtime; it does not create a separate lifecycle or
-disable managed delivery.
+There is one canonical Channel host: [`server.ts`](./server.ts).
+[`app/index.ts`](./app/index.ts) composes one `CopilotKitIntelligence`, one
+`CopilotRuntime`, and two adapter-free managed Channels: the configured base
+name for Slack and `<base>-teams` for Microsoft Teams. Intelligence owns all
+platform credentials and attachments.
 
 The launch pins the requested canaries:
 
@@ -55,15 +53,7 @@ Prerequisites: Node.js 22+, pnpm, Python 3.12, and
    cd agent && uv sync && cd ..
    ```
 
-2. Configure the Python agent.
-
-   ```bash
-   cp agent/.env.example agent/.env
-   ```
-
-   Set `OPENAI_API_KEY`. Tavily, Linear, and Notion are optional.
-
-3. Configure the Channel host.
+2. Configure both services from the shared root environment.
 
    ```bash
    cp .env.example .env
@@ -72,18 +62,23 @@ Prerequisites: Node.js 22+, pnpm, Python 3.12, and
    Set:
 
    ```dotenv
+   OPENAI_API_KEY=sk-...
    AGENT_URL=http://localhost:8123/
    INTELLIGENCE_API_KEY=cpk-...
    ```
+
+   Tavily, Linear, and Notion are optional. Both Node and Python load this root
+   `.env`; Railway supplies the same values as service variables.
 
    `INTELLIGENCE_API_URL` and `INTELLIGENCE_GATEWAY_WS_URL` default to the
    production Intelligence endpoints. `INTELLIGENCE_CHANNEL_NAME` defaults to
    `opentag`.
 
-4. In Intelligence, create the Channel that matches
-   `INTELLIGENCE_CHANNEL_NAME` and attach managed Slack and/or Microsoft Teams.
+3. In Intelligence, create two managed Channels: `opentag` for Slack and
+   `opentag-teams` for Microsoft Teams. If you change the base name, use that
+   name and `<base>-teams`.
 
-5. Run both services.
+4. Run both services.
 
    ```bash
    pnpm agent
@@ -110,34 +105,18 @@ in [`agent/agent.py`](./agent/agent.py); the Channel UI and behavior live under
 
 ## Platform setup
 
-### Intelligence (recommended)
+### Intelligence
 
 Use the normal Intelligence flow for both launch platforms:
 
-1. Create one OpenTag project and Channel in Intelligence.
-2. Issue a runtime API key.
-3. Configure the Slack and/or Microsoft Teams attachment in Intelligence.
-4. Run one `pnpm channel` process with that Channel name and key.
+1. Create one OpenTag project.
+2. Create the base Channel for Slack and `<base>-teams` for Microsoft Teams.
+3. Issue a runtime API key.
+4. Configure each platform attachment on its matching Channel.
+5. Run one `pnpm channel` process with the base Channel name and key.
 
 No organization, project, Channel ID, or runtime-instance ID environment
-variables are required.
-
-### Optional direct adapters
-
-For local development, set either complete pair in the root `.env`:
-
-```dotenv
-SLACK_BOT_TOKEN=xoxb-...
-SLACK_APP_TOKEN=xapp-...
-
-TEAMS_CLIENT_ID=...
-TEAMS_CLIENT_SECRET=...
-```
-
-`TEAMS_TENANT_ID` and `TEAMS_PORT` are optional. Incomplete credential pairs
-fail at startup instead of silently disabling an adapter. The Intelligence
-connection remains the owner of the runtime, while each direct adapter stays
-isolated from the managed Channel.
+variables—or Slack/Teams credentials—are required by the runtime.
 
 ### Slack manifests
 
@@ -155,9 +134,9 @@ Slack app**. Reusing it preserves the bot user, workspace installation, and
 - `TAVILY_API_KEY` enables live web research. Without it, OpenTag still chats,
   plans, uses its virtual filesystem, and renders UI from model knowledge.
 - `LINEAR_API_KEY` enables the hosted Linear MCP.
-- Notion can be used locally with `pnpm notion-mcp`; copy the same
-  `NOTION_MCP_AUTH_TOKEN` into the root `.env` and `agent/.env`. The Railway
-  launch intentionally has no Notion sidecar.
+- Notion can be used locally with `pnpm notion-mcp`; its token and shared
+  `NOTION_MCP_AUTH_TOKEN` live in the root `.env`. The Railway launch
+  intentionally has no Notion sidecar.
 
 Every Linear and Notion mutation is intercepted in code before the MCP request
 runs. The interceptor emits `confirm_write` and proceeds only after approval;
@@ -174,10 +153,11 @@ both sourced from `CopilotKit/OpenTag` on `main`:
 | `channel` | repository root | `pnpm channel` | Channels readiness before listen |
 
 The Channel reaches the agent over Railway private networking. Railway sets
-`INTELLIGENCE_CHANNEL_NAME=kite` for the production cutover and preserves the
-runtime API key. OpenAI is required on `agent`; Tavily and Linear secrets are
-optional. Connecting both services to `main` enables GitHub-triggered
-deployments after merges.
+`INTELLIGENCE_CHANNEL_NAME=kite`, so the runtime declares the managed `kite`
+Slack Channel and `kite-teams` Teams Channel. It preserves the runtime API key.
+OpenAI is required on `agent`; Tavily and Linear secrets are optional.
+Connecting both services to `main` enables GitHub-triggered deployments after
+merges.
 
 The repository configuration does not mutate the existing production Railway
 project. Inventory and cutover should happen after Railway authentication.
