--- agent/.env.example
@@ -1,26 +1,0 @@
-# OpenTag deep-research agent (agent/) — served over AG-UI on SERVER_PORT.
-# The OpenTag Channel host points AGENT_URL at http://localhost:8123/
-# Only OPENAI_API_KEY is required; the agent chats + generates UI components without Tavily.
-OPENAI_API_KEY=sk-...
-# OPTIONAL — web research. Without it the agent still works (chat + UI components,
-# answers from its own knowledge); with it, live web research turns on. https://tavily.com
-# TAVILY_API_KEY=tvly-...
-# OpenAI model (defaults to gpt-5.5, matching the rest of OpenTag).
-# OPENAI_MODEL=gpt-5.5
-# Server bind (defaults 0.0.0.0:8123).
-# SERVER_HOST=0.0.0.0
-# SERVER_PORT=8123
-
-# ── Internal research sources (optional) ────────────────────────────────
-# Setting LINEAR_API_KEY and/or NOTION_MCP_AUTH_TOKEN lets the agent research
-# the team's Linear issues / Notion docs alongside (or instead of) the web —
-# see internal_source_tools() in tools.py. Leave both blank to research the
-# web only.
-# Hosted Linear MCP bearer token (linear.app → Settings → API → Personal API
-# keys). LINEAR_MCP_URL defaults to the hosted Linear MCP if unset.
-# LINEAR_API_KEY=lin_api_...
-# LINEAR_MCP_URL=https://mcp.linear.app/mcp
-# Bearer token for the local Notion MCP sidecar (`pnpm notion-mcp`).
-# NOTION_MCP_URL defaults to that local sidecar if unset.
-# NOTION_MCP_AUTH_TOKEN=choose-a-strong-shared-secret
-# NOTION_MCP_URL=http://127.0.0.1:3001/mcp
