+++ agent/tests/test_write_confirmation.py
@@ -1,0 +1,184 @@
+import asyncio
+
+import copilotkit.langgraph
+import pytest
+import write_confirmation
+from langchain_core.tools import StructuredTool
+from langchain_mcp_adapters.interceptors import MCPToolCallRequest
+
+
+def test_write_confirmation_emits_the_copilotkit_interrupt_envelope(monkeypatch):
+    calls = []
+    handler_calls = []
+
+    def fake_interrupt(value):
+        calls.append(value)
+        return {"confirmed": True}
+
+    monkeypatch.setattr(copilotkit.langgraph, "interrupt", fake_interrupt)
+
+    async def handler(request):
+        handler_calls.append(request)
+        return "write-result"
+
+    request = MCPToolCallRequest(
+        name="create_issue",
+        args={"title": "Checkout 500s"},
+        server_name="linear",
+    )
+    result = asyncio.run(
+        write_confirmation.WriteConfirmationInterceptor()(request, handler)
+    )
+
+    assert result == "write-result"
+    assert handler_calls == [request]
+    assert len(calls) == 1
+    assert set(calls[0]) == {
+        "__copilotkit_interrupt_value__",
+        "__copilotkit_messages__",
+    }
+    assert calls[0]["__copilotkit_interrupt_value__"] == {
+        "action": "confirm_write",
+        "args": {
+            "action": "Create issue",
+            "detail": '{"title": "Checkout 500s"}',
+        },
+    }
+
+
+def test_write_confirmation_interceptor_leaves_annotated_reads_unguarded(
+    monkeypatch,
+):
+    interrupt_calls = []
+    handler_calls = []
+
+    async def read_issue(issue_id: str):
+        return issue_id
+
+    read_tool = StructuredTool.from_function(
+        coroutine=read_issue,
+        name="get_issue",
+        description="Read an issue",
+        metadata={"readOnlyHint": True},
+    )
+    interceptor = write_confirmation.WriteConfirmationInterceptor()
+    interceptor.register_tools([read_tool])
+    monkeypatch.setattr(
+        write_confirmation,
+        "copilotkit_interrupt",
+        lambda **kwargs: interrupt_calls.append(kwargs),
+    )
+
+    async def handler(request):
+        handler_calls.append(request)
+        return "read-result"
+
+    result = asyncio.run(
+        interceptor(
+            MCPToolCallRequest(
+                name="get_issue",
+                args={"issue_id": "CPK-9"},
+                server_name="linear",
+            ),
+            handler,
+        )
+    )
+
+    assert result == "read-result"
+    assert len(handler_calls) == 1
+    assert interrupt_calls == []
+
+
+def test_write_confirmation_interceptor_blocks_a_declined_mutation(monkeypatch):
+    interrupt_calls = []
+    handler_calls = []
+    interceptor = write_confirmation.WriteConfirmationInterceptor()
+
+    def decline(**kwargs):
+        interrupt_calls.append(kwargs)
+        return '{"confirmed": false}', {"confirmed": False}
+
+    monkeypatch.setattr(write_confirmation, "copilotkit_interrupt", decline)
+
+    async def handler(request):
+        handler_calls.append(request)
+        return "write-result"
+
+    result = asyncio.run(
+        interceptor(
+            MCPToolCallRequest(
+                name="create_issue",
+                args={"title": "Checkout 500s"},
+                server_name="linear",
+            ),
+            handler,
+        )
+    )
+
+    assert handler_calls == []
+    assert interrupt_calls == [
+        {
+            "action": "confirm_write",
+            "args": {
+                "action": "Create issue",
+                "detail": '{"title": "Checkout 500s"}',
+            },
+        }
+    ]
+    assert result.content[0].text == (
+        "Write cancelled by the user; no changes were made."
+    )
+
+
+def test_write_confirmation_interceptor_runs_an_approved_mutation(monkeypatch):
+    handler_calls = []
+    interceptor = write_confirmation.WriteConfirmationInterceptor()
+    monkeypatch.setattr(
+        write_confirmation,
+        "copilotkit_interrupt",
+        lambda **_kwargs: ('{"confirmed": true}', {"confirmed": True}),
+    )
+
+    async def handler(request):
+        handler_calls.append(request)
+        return "write-result"
+
+    request = MCPToolCallRequest(
+        name="update_issue",
+        args={"id": "CPK-9", "title": "Checkout 500s"},
+        server_name="linear",
+    )
+    result = asyncio.run(interceptor(request, handler))
+
+    assert result == "write-result"
+    assert handler_calls == [request]
+
+
+def test_write_confirmation_interceptor_rejects_a_malformed_resume(
+    monkeypatch,
+):
+    handler_calls = []
+    interceptor = write_confirmation.WriteConfirmationInterceptor()
+    monkeypatch.setattr(
+        write_confirmation,
+        "copilotkit_interrupt",
+        lambda **_kwargs: ("unexpected", {"unexpected": True}),
+    )
+
+    async def handler(request):
+        handler_calls.append(request)
+        return "write-result"
+
+    with pytest.raises(RuntimeError, match="confirmed"):
+        asyncio.run(
+            interceptor(
+                MCPToolCallRequest(
+                    name="create_issue",
+                    args={"title": "Checkout 500s"},
+                    server_name="linear",
+                ),
+                handler,
+            )
+        )
+
+    assert handler_calls == []
