+++ agent/tools.py
@@ -9,138 +9,11 @@ to prevent subagent text from leaking to the frontend via LangChain callback pro
 """
 
 import os
-import asyncio
-import json
-import logging
-from datetime import datetime, timezone
 from typing import Any
 from concurrent.futures import ThreadPoolExecutor
-from langchain_core.tools import BaseTool, tool
+from langchain_core.tools import tool
 from langchain_core.messages import HumanMessage
 from tavily import TavilyClient
-from langchain_mcp_adapters.client import MultiServerMCPClient
-from langchain_mcp_adapters.interceptors import (
-    MCPToolCallRequest,
-    MCPToolCallResult,
-)
-from copilotkit.langgraph import copilotkit_interrupt
-from mcp.types import CallToolResult, TextContent
-
-
-logger = logging.getLogger(__name__)
-
-
-def _emit_internal_source_error(
-    error: BaseException,
-    *,
-    message: str,
-    source: str,
-    recovery: str,
-) -> None:
-    """Emit an observable error without logging credentials or server details."""
-    logger.error(
-        {
-            "error": message,
-            "context": {
-                "component": "internal_source_tools",
-                "exception_type": type(error).__name__,
-                "recovery": recovery,
-                "source": source,
-            },
-            "timestamp": datetime.now(timezone.utc).isoformat(),
-        }
-    )
-
-
-@tool
-def confirm_write(action: str, detail: str | None = None):
-    """Pause before a Linear or Notion write and ask the user to approve it.
-
-    Args:
-        action: One-line summary of exactly what will be written.
-        detail: Optional drafted title, description, or outline.
-
-    Returns:
-        The user's resume response after the confirmation interrupt.
-    """
-    answer, _response = copilotkit_interrupt(
-        action="confirm_write",
-        args={"action": action, "detail": detail},
-    )
-    return answer
-
-
-class WriteConfirmationInterceptor:
-    """Enforce confirmation inside every mutating internal-source tool call.
-
-    MCP tool annotations are the authority for read-only calls. Anything not
-    explicitly marked read-only fails closed and pauses before the remote tool
-    handler can run. The two Notion search operations below use POST while
-    remaining read-only, so its OpenAPI bridge cannot infer their annotation.
-    """
-
-    _KNOWN_READ_ONLY_TOOLS = {
-        "API-post-search",
-        "API-query-data-source",
-    }
-
-    def __init__(self):
-        self._read_only_tools = set(self._KNOWN_READ_ONLY_TOOLS)
-
-    def register_tools(self, tools: list[BaseTool]) -> None:
-        for source_tool in tools:
-            metadata = source_tool.metadata or {}
-            if metadata.get("readOnlyHint") is True:
-                self._read_only_tools.add(source_tool.name)
-
-    async def __call__(
-        self,
-        request: MCPToolCallRequest,
-        handler,
-    ) -> MCPToolCallResult:
-        if request.name in self._read_only_tools:
-            return await handler(request)
-
-        action = request.name.replace("_", " ").replace("-", " ").strip()
-        action = action[:1].upper() + action[1:]
-        detail = json.dumps(
-            request.args,
-            ensure_ascii=False,
-            sort_keys=True,
-            default=str,
-        )
-        _answer, response = copilotkit_interrupt(
-            action="confirm_write",
-            args={"action": action, "detail": detail},
-        )
-
-        if isinstance(response, str):
-            try:
-                response = json.loads(response)
-            except json.JSONDecodeError:
-                response = None
-
-        if (
-            not isinstance(response, dict)
-            or not isinstance(response.get("confirmed"), bool)
-        ):
-            raise RuntimeError(
-                "confirm_write resume must contain a boolean `confirmed` value"
-            )
-
-        if response["confirmed"] is False:
-            return CallToolResult(
-                content=[
-                    TextContent(
-                        type="text",
-                        text=(
-                            "Write cancelled by the user; no changes were made."
-                        ),
-                    )
-                ]
-            )
-
-        return await handler(request)
 
 
 def _do_internet_search(query: str, max_results: int = 5) -> list[dict[str, Any]]:
@@ -289,104 +162,3 @@ Rules:
 
     print(f"[TOOL] research: completed with {len(result['sources'])} sources")
     return result
-
-
-def internal_source_tools() -> list:
-    """Notion + Linear MCP tools, included only when their env is present.
-
-    Reuses OpenTag's Phase-1 MCP servers so the agent can research the team's
-    own docs/issues alongside the web. Linear uses its hosted MCP server
-    (`LINEAR_API_KEY`, with `LINEAR_MCP_URL` defaulting to `mcp.linear.app`);
-    Notion uses the optional local MCP sidecar (`NOTION_MCP_AUTH_TOKEN`, with
-    `NOTION_MCP_URL` defaulting to localhost). Each server is independently
-    optional.
-
-    `langchain_mcp_adapters`'s `MultiServerMCPClient.get_tools()` is
-    async-only; this function runs it to completion on a short-lived event
-    loop (`asyncio.run`) so it can be called synchronously from
-    `build_agent()`. Each server is connected individually so one server
-    failing to load never drops the other's tools. These integrations are
-    optional deployment enhancements, so a down or misconfigured server emits
-    a structured error event and is skipped instead of blocking agent startup.
-
-    Returns:
-        list: LangChain tools from the configured MCP server(s), or an empty
-            list when neither integration is configured (or every configured
-            optional integration emitted an error and was skipped).
-    """
-    connections: dict[str, dict[str, Any]] = {}
-
-    linear_api_key = os.environ.get("LINEAR_API_KEY")
-    if linear_api_key:
-        connections["linear"] = {
-            "transport": "streamable_http",
-            "url": os.environ.get("LINEAR_MCP_URL", "https://mcp.linear.app/mcp"),
-            "headers": {"Authorization": f"Bearer {linear_api_key}"},
-        }
-
-    notion_auth_token = os.environ.get("NOTION_MCP_AUTH_TOKEN")
-    if notion_auth_token:
-        connections["notion"] = {
-            "transport": "streamable_http",
-            "url": os.environ.get("NOTION_MCP_URL", "http://127.0.0.1:3001/mcp"),
-            "headers": {"Authorization": f"Bearer {notion_auth_token}"},
-        }
-
-    if not connections:
-        return []
-
-    async def _load_all() -> list:
-        loaded: list = []
-        for name, connection in connections.items():
-            try:
-                confirmation = WriteConfirmationInterceptor()
-                server_client = MultiServerMCPClient(
-                    {name: connection},
-                    tool_interceptors=[confirmation],
-                )
-                # Bound connect time so a hung MCP server (accepts the
-                # socket but never responds) can't stall build_agent() -
-                # and thus the whole app, including /health - at import
-                # time.
-                server_tools = await asyncio.wait_for(
-                    server_client.get_tools(), timeout=8.0
-                )
-                confirmation.register_tools(server_tools)
-                loaded.extend(server_tools)
-                print(
-                    f"[TOOLS] internal_source_tools: loaded {len(server_tools)} "
-                    f"tool(s) from {name}"
-                )
-            except asyncio.TimeoutError as error:
-                # Optional MCP integrations must not block the core agent, but
-                # a configured source failing is distinct from an intentionally
-                # disabled source and must remain observable.
-                _emit_internal_source_error(
-                    error,
-                    message="Configured MCP server timed out while loading tools",
-                    source=name,
-                    recovery="skip_optional_integration",
-                )
-            except Exception as error:
-                _emit_internal_source_error(
-                    error,
-                    message=(
-                        "Configured MCP server was unavailable while loading tools"
-                    ),
-                    source=name,
-                    recovery="skip_optional_integration",
-                )
-        return loaded
-
-    try:
-        return asyncio.run(_load_all())
-    except Exception as error:
-        # Loading optional integrations is recoverable for the core agent, but
-        # event-loop failures still surface as structured operational errors.
-        _emit_internal_source_error(
-            error,
-            message="Failed to initialize MCP tool loading",
-            source="all",
-            recovery="skip_optional_integrations",
-        )
-        return []
