+++ app/components/index.ts
@@ -1,10 +1,9 @@
 /**
  * App-specific render components — agent-renderable Block Kit cards authored
- * with the `@copilotkit/channels` JSX vocabulary.
+ * with the umbrella `@copilotkit/channels` package's JSX vocabulary.
  *
  * Each component is a plain `ComponentFn` returning a `<Message>` tree; its
- * exported zod prop schema doubles as the render-tool input schema. Render a
- * component with `renderSlackMessage(renderToIR(<IssueCard {...props} />))`.
+ * exported zod prop schema doubles as the render-tool input schema.
  */
 export { IssueCard, issueCardSchema } from "./issue-card.js";
 export type { IssueCardProps } from "./issue-card.js";
