+++ app/sender-context.ts
@@ -5,7 +5,7 @@ import type { ContextEntry, PlatformUser } from "@copilotkit/channels";
  * "as" them (filter Linear by their email, @-mention/tag them). The platform
  * adapter resolves `{ id, name?, email? }` per turn; if it's absent there's
  * nothing to attribute, so we add no entry. `platform` is the source surface,
- * so the label is correct for managed and direct Slack/Teams turns alike.
+ * so the label is correct for managed Slack and Teams turns alike.
  */
 export function senderContext(
   user: PlatformUser | undefined,
